tentativa de migrar para hexagonal
