+++ package.json
@@ -6,66 +6,67 @@
     "private": true,
     "license": "UNLICENSED",
     "scripts": {
-      "prebuild": "rimraf dist",
-      "build": "nest build",
-      "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-      "start": "nest start",
-      "start:dev": "nest start --watch",
-      "start:debug": "nest start --debug --watch",
-      "start:prod": "node dist/main",
-      "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
-      "test": "jest",
-      "test:watch": "jest --watch",
-      "test:cov": "jest --coverage",
-      "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
-      "test:e2e": "jest --config ./test/jest-e2e.json"
+        "prebuild": "rimraf dist",
+        "build": "nest build",
+        "format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
+        "start": "nest start",
+        "start:dev": "nest start --watch",
+        "start:debug": "nest start --debug --watch",
+        "start:prod": "node dist/main",
+        "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
+        "test": "jest",
+        "test:watch": "jest --watch",
+        "test:cov": "jest --coverage",
+        "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
+        "test:e2e": "jest --config ./test/jest-e2e.json"
     },
     "dependencies": {
-      "@nestjs/common": "^9.0.0",
-      "@nestjs/core": "^9.0.0",
-      "@nestjs/platform-express": "^9.0.0",
-      "reflect-metadata": "^0.1.13",
-      "rimraf": "^3.0.2",
-      "rxjs": "^7.2.0"
+        "@nestjs/common": "^9.0.0",
+        "@nestjs/core": "^9.0.0",
+        "@nestjs/platform-express": "^9.0.0",
+        "axios": "^1.7.4",
+        "reflect-metadata": "^0.1.13",
+        "rimraf": "^3.0.2",
+        "rxjs": "^7.2.0"
     },
     "devDependencies": {
-      "@nestjs/cli": "^9.0.0",
-      "@nestjs/schematics": "^9.0.0",
-      "@nestjs/testing": "^9.0.0",
-      "@types/express": "^4.17.13",
-      "@types/jest": "28.1.8",
-      "@types/node": "^16.0.0",
-      "@types/supertest": "^2.0.11",
-      "@typescript-eslint/eslint-plugin": "^5.0.0",
-      "@typescript-eslint/parser": "^5.0.0",
-      "eslint": "^8.0.1",
-      "eslint-config-prettier": "^8.3.0",
-      "eslint-plugin-prettier": "^4.0.0",
-      "jest": "28.1.3",
-      "prettier": "^2.3.2",
-      "source-map-support": "^0.5.20",
-      "supertest": "^6.1.3",
-      "ts-jest": "28.0.8",
-      "ts-loader": "^9.2.3",
-      "ts-node": "^10.0.0",
-      "tsconfig-paths": "4.1.0",
-      "typescript": "^4.7.4"
+        "@nestjs/cli": "^9.0.0",
+        "@nestjs/schematics": "^9.0.0",
+        "@nestjs/testing": "^9.0.0",
+        "@types/express": "^4.17.13",
+        "@types/jest": "28.1.8",
+        "@types/node": "^16.0.0",
+        "@types/supertest": "^2.0.11",
+        "@typescript-eslint/eslint-plugin": "^5.0.0",
+        "@typescript-eslint/parser": "^5.0.0",
+        "eslint": "^8.0.1",
+        "eslint-config-prettier": "^8.3.0",
+        "eslint-plugin-prettier": "^4.0.0",
+        "jest": "28.1.3",
+        "prettier": "^2.3.2",
+        "source-map-support": "^0.5.20",
+        "supertest": "^6.1.3",
+        "ts-jest": "28.0.8",
+        "ts-loader": "^9.2.3",
+        "ts-node": "^10.0.0",
+        "tsconfig-paths": "4.1.0",
+        "typescript": "^4.7.4"
     },
     "jest": {
-      "moduleFileExtensions": [
-        "js",
-        "json",
-        "ts"
-      ],
-      "rootDir": "src",
-      "testRegex": ".*\\.spec\\.ts$",
-      "transform": {
-        "^.+\\.(t|j)s$": "ts-jest"
-      },
-      "collectCoverageFrom": [
-        "**/*.(t|j)s"
-      ],
-      "coverageDirectory": "../coverage",
-      "testEnvironment": "node"
+        "moduleFileExtensions": [
+            "js",
+            "json",
+            "ts"
+        ],
+        "rootDir": "src",
+        "testRegex": ".*\\.spec\\.ts$",
+        "transform": {
+            "^.+\\.(t|j)s$": "ts-jest"
+        },
+        "collectCoverageFrom": [
+            "**/*.(t|j)s"
+        ],
+        "coverageDirectory": "../coverage",
+        "testEnvironment": "node"
     }
-  }
+}
