+++ src/conta-corrente/domain/ports/geolocation.port.ts
@@ -1,0 +1,3 @@
+export class GeolocationPort{
+    
+}
