+++ src/conta-poupanca/conta-poupanca.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { ContaPoupancaController } from './conta-poupanca.controller';
-import { ContaPoupancaService } from './conta-poupanca.service';
+import { ContaPoupancaController } from './adapters/inbound/conta-poupanca.controller';
+import { ContaPoupancaService } from './application/conta-poupanca.service';
 @Module({
   controllers: [ContaPoupancaController],
   providers: [ContaPoupancaService],
